DIM v17r5
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@97910 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: ced68f0ee2394b118ac262470ffecc1950af8f8c

+++ Online/Online/DIM/dim/dim_common.h
@@ -46,7 +46,7 @@ typedef enum { SRC_NONE, SRC_DIS, SRC_DIC, SRC_DNS, SRC_DNA, SRC_USR }SRC_TYPES;
 typedef __int64		longlong;
 #else
 #include <sys/types.h> 
-typedef int64_t	longlong;
+typedef long long int	longlong;
 #endif
 
 #endif
